displays the current list of scheduled tasks.
The Task Scheduler Grid View displays the current list of scheduled tasks. Click the Create Scheduled Tasks button to open the New Taskdialog box, where you can add information about the new task and schedule the notification emails. In the Task tab, enter a name for the new task in the Name field.